Update all operations for scale and toughtness

+++ vm/srcs/opp/long_fork.c
@@ -26,7 +26,7 @@ int		long_fork(t_mars *mars, t_processus *process)
 	unsigned	dest;
 
 	dest = ft_get_mars_value(mars, process->pc + 1, IND_SIZE);
-	ft_copy_process(process, mars, ft_global_restriction(process->pc + dest));
+	ft_copy_process(process, mars, ft_global_restriction(process->pc + (short)dest));
 	process->bytes_to_jump = process->pc + IND_SIZE;
 	return (SUCCESS);
 }
